# storage/engine: reimplement MVCCDeleteRange in terms of MVCCScan