@bdarnell 2018-01-17
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 2 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21507#-:-L34VEV3-Ntiq1lSdkok:bnfp4nl)*
<!-- Sent from Reviewable.io -->